FishLight_PCB: First rev of schematic and board
These are the preliminary commits of the FishLight_PCB schematic and
layout.  They are rudimentary and essentially just set up the
mechanicals for the board to fit nicely onto the heatsink,  Also put
down the Arduino and RTC because they are rather large.  I'm not sure
what the .b#* files are but they seem to be autogenerated.  So
committing them as well.