FishLight_PCB: Added CD74HC4067 16x MUX
Added the CD74HC4067 16-Channel MUX.  This is going to control which
digital pot gets written to since they all have the same I2C address.  I
hope it will be fast enough to write a single register on each pot.
Pulled the SSOP20 package from Sparkfun and modified it for a 24 pin
package.  Hooked up the SDA signals on 3 of the pots and the Arduino.
Still need to connect the Digital control signals S0-3.  Need to choose
which pins those will be.