FishLight_PCB: Added 5V regulator and 3 pin Molex
Added the L7805 5V regulator.  This may not be necessary unless I want
to use the Spark Core.  The Spark has a max input voltage of 6V, making
the 12V regulator unnecessary.  Perhaps I will pick a regulator that can
be changed to 5V or 12V depending on what I want to use just by changing
one set resistor.  Also added a few 3 pin Molex connectors that I can
use for temperature sensors.  I haven't picked a part for that yet but I
may go with the DS18B20 even tho it is kind of expensive.  But it is
really nice since it has an Arduino library already and it only uses 1
wire to talk to the Arduino and it can have a large number of sensors on
the same wire since they all have different pre-programmed addresses.