FishLight_PCB: Upgrade to Spark Core
I am planning on using the Spark Core for this project so that I can
have a wireless connection to program the board and get updates on
things like temperature wirelessly.  Added the Spark Core parts library,
the Adafruit parts library, updated a few of my parts and changed to the
Spark Core.