FishLight_PCB: Added fans and moved parts around
Moved parts around on the board to be a bit more friendly to the 24V
rail.  Also changed the color of the top silk screen layer but I doubt
that will stick.  Added some fan connectors in the hopes that I can add
a 3 pin fan in the future.  Need to come up with a good control circuit
for that tho.  I would like to have speed control not just on and off.