FishLight_PCB: Fixed fan transistors
Fixed the fan connector transistor pull down resistors.  These were
connected to drain but they need to be connected to gate so that the
transistor shuts off correctly.  Also added some Solid State Relays to
possibly turn on and off some AC loads (such as a pump or the heater)
for emergency situations.  These will probably go away though.  Kind of
an interesting idea is all.