FishLight_PCB: Add NC7SZ373 D-Latch
The NC7SZ373 is a D-Latch which can be used to turn on and off a device
with a transistor attached by setting 'D" to on/off and then pulsing the
'LE' input.  This will require 2 pins, D and LE and 2 muxes which
correspond to these inputs.  Possibly only 1 mux if I use less than 7
devices on a 4 - 15 mux.