FishLight_PCB: Added LED strings 4 - 8
Added LED strings 4 - 8 and removed the 'green' string.  This will allow
me to control 2 tanks or one really big tank in the future.  Added all
of the accompanying parts and moved a lot of stuff around to make it all
fit on the schematic.  Still need to lay them out on the board tho.