FishLight_PCB: Replaced the RTC and added VIAs
Replaced the giant RTC breakout board with the new DS3231M RTC.  This
can fit right under the Spark Core so it takes up very little space.
The battery is still kinda big but its not a big deal.  Also added VIAs
to the CAT4101s that are on the top side and to the I-Sense resistors
that are new.