FishLight_PCB: Shrink schematic symbols
In an effort to save space, I have shrunk many of my own schematic
symbols by making pins small instead of medium in their properties.
This saved a lot of space in the schematic.  I also added 9 2pin screw
terminals for 12V devices and changed the current limiting resistor on
the fans to limit current to the transistor, not the fan itself.  I may
also put some LEDs on the 12V devices and the SSRs in order to show what
devices are on or off.  May also want to add another SSR.