FishLight_PCB: Add second SCH frame and replace 12V reg
Added a second Schematic frame and placed it in a place that will allow
it to be printed correctly.  This will open up a lot of room for more
devices on my schematic, specifically my 12V devices.  Also changed my
12V regulator to a higher current (3A) version.  This will allow me to
have more 12V devices.  Also added 3 more, total 5, SSRs.  This may be
too many for my case but it will be interesting to try it out.