FishLight_PCB: Fix Spark pin assignments
The Spark Core schematic symbol is a bit misleading.  The RAW pin is
actually VIN, VDDA is the 'clean' 3.3V output from the voltage
regulator, and VDD is actually the 3.3V pin, which could power the Core
if I was using a 3.3V regulator.  Also, all analog pins are PWM pins and
the D0/D1 pins are SDA/SCL.  I also fixed the CAT4101 connections to
connect to actual PWM pins.