FishLight_PCB: Started work on 12V devices
The 12V devices require the ability to turn them on and off but also to
hold them on or off without constant need of the Spark Core to supply a
transistor.  I am going to use a MUX with a D Flip-Flop and a transistor
to hold the devices High or Low and then go and do other things while
the FF holds their state.  I have placed the MUXes, the beginnings of
the FF circuit, and the terminals.  I have also connected the Spark
outputs D2 - D5 to the MUX S inputs.  I also changed the TMP sensors to
D6.