FishLight_PCB: Rearranged some parts
Rearranged the CAT4101s to make more room.  Also changed their screw
terminal assignments to limit the number of crossings.  This will have
to be resolved in software now because the screw terminal assignments
are all screwed around.  Also changed around the Spark to make more room
for the AC SSRs.  These are now placed on the board.  Also moved around
the 12V devices screw terminals and circuitry to limit MUX crossings.
This is going to be super difficult to route and may need to reduce the
number of 12V devices.