FishLight_PCB: Lots of changes
Updated the package for the AD5248 digital potentiometers so that the
pins do not cause DRC errors.  Also continued routing and moved around
the LED ground connections so that they would be easier to use.  Also
changed the address assignments for the INA219 current sensors so that
they would be easier to route.  Did a DRC check and approved a number of
overlaps.  I also made a few attempts at putting in flyback diodes but
they are too big for my small design.  I may need to revisit this.  Also
looked into fuses for the AC devices but I did not find any designs at
Sparkfun that used fuses on their SSRs so I can probably skip this for
now.  Again, may need to revisit this.