FishLight_PCB: Remove the RTC
The RTC isn't really needed because the Spark Core can get the time
directly from the internet and it has an onboard RTC that can be used
instead.  The only downside is that it doesn't have a battery but it can
just sync with the internet at startup.  It also has EEPROM like memory
that can be accessed to store data as needed, just like an RTC would
have.