FishLight_PCB: Lots and lots of routing
Most of the routing is done now.  Got the MUX signals mostly figured out
but there are still a few that are going to be very tough, like the AC
devices and the fans.  The AC devices still need to be hooked to their
terminals too and this worries me because there isn't much space to
isolate the lines.  I guess the good news is that if it shorts that just
means it gets connected and it wont be controllable from the Spark.  But
this isn't a good thing either.